Seperated things + plus attendance attempt
